# [viewer] speed up dedup interface