--- a/ymir/backend/src/ymir_viewer/common/constants/config.go
+++ b/ymir/backend/src/ymir_viewer/common/constants/config.go
@@ -15,9 +15,9 @@ type Config struct {
 	ViewerPort int
 	ViewerURI  string
 
-	MongoDBURI     string
-	MongoDBName    string
-	MongoDBNoCache bool
+	MongoDBURI       string
+	MongoDataDBName  string
+	MongoDataDBCache bool
 
 	InnerTimeout time.Duration
 }