--- a/ymir/backend/src/ymir_viewer/config.yml
+++ b/ymir/backend/src/ymir_viewer/config.yml
@@ -6,7 +6,7 @@ ViewerPort: 9527
 ViewerURI: ""
 
 MongoDBURI: "mongodb://127.0.0.1:27017"
-MongoDBName: "YMIR"
-MongoDBNoCache: true
+MongoDataDBName: "YMIR-data"
+MongoDataDBCache: true
 
 InnerTimeout: 61 # +1 for elder.