--- a/ymir/backend/src/ymir_viewer/services/mongo.go
+++ b/ymir/backend/src/ymir_viewer/services/mongo.go
@@ -436,3 +436,55 @@ func (s *MongoServer) QueryDatasetStats(
 
 	return queryData
 }
+
+func (s *MongoServer) QueryDatasetDup(
+	mirRepo0 *constants.MirRepo,
+	mirRepo1 *constants.MirRepo,
+) *constants.QueryDatasetDupResult {
+	defer tools.TimeTrack(time.Now())
+
+	collection0, collectionName0 := s.getRepoCollection(mirRepo0)
+	totalCount0, err := collection0.CountDocuments(s.Ctx, bson.M{}, &options.CountOptions{})
+	if err != nil {
+		panic(err)
+	}
+
+	collection1, collectionName1 := s.getRepoCollection(mirRepo1)
+	totalCount1, err := collection0.CountDocuments(s.Ctx, bson.M{}, &options.CountOptions{})
+	if err != nil {
+		panic(err)
+	}
+
+	// No-SQL Aggregate Trick: always set smaller dataset to be joined.
+	collectionJoined := collection0
+	collectionNameToJoin := collectionName1
+	if totalCount0 > totalCount1 {
+		collectionJoined = collection1
+		collectionNameToJoin = collectionName0
+	}
+
+	lookupStage := bson.D{
+		bson.E{
+			Key: "$lookup",
+			Value: bson.D{
+				bson.E{Key: "from", Value: collectionNameToJoin},
+				bson.E{Key: "localField", Value: "asset_id"},
+				bson.E{Key: "foreignField", Value: "asset_id"},
+				bson.E{Key: "as", Value: "joinAssets"},
+			},
+		},
+	}
+	showLoadedCursor, err := collectionJoined.Aggregate(s.Ctx, mongo.Pipeline{lookupStage})
+	if err != nil {
+		panic(err)
+	}
+	var showsLoaded []bson.M
+	if err = showLoadedCursor.All(s.Ctx, &showsLoaded); err != nil {
+		panic(err)
+	}
+
+	return &constants.QueryDatasetDupResult{
+		Duplication: len(showsLoaded),
+		TotalCount:  map[string]int64{mirRepo0.BranchID: totalCount0, mirRepo1.BranchID: totalCount1},
+	}
+}