Background: 
最初设计dedup接口时，是有两个假设：

1. 两个数据集已有缓存（比如在之前做过可视化）
2. 对已有缓存的join查询会很快：1-2s

实际测试中发现这两个假设并不成立：

1. 缓存大多数情况不存在，此时这个接口需要即时对不存在的数据集先做缓存。缓存完成后才能查询，性能很差（10万数据集读取+缓存时间在单库55s左右）
2. 外接表查询时间慢，10万+5万数据需要28s左右

benchmark（10万数据集）：

- 读取metadata（3.4MB）：0.38s
- 读取metadata+annotation（3.4MB+40MB）：6s
- 数据格式转换：25s
- mongodb入库：8s
- mongodb索引：16S
- 完整缓存流程：6+25+8+16=55s

优化策略：
由于dedup只需要metadata，而读取metadata非常快。修改为直接从metadata中读取两个库的asset id并去重。同时在后台做缓存

性能影响（10万+5万双库估计）
优化前：

-   缓存都存在时：28s
-   只存在一个缓存时：55s+28s or 27s+28s
-   缓存都不存在时：55s+27s+28s

优化后：

- 0.38s+0.17s+0.035s=0.58s